restored route grouping for easier layouts
