+++ app/(public)/[...slug]/metadata.tsx
