+++ app/(public)/search/results.tsx
